Replaced interface AssociationCascader with implementation
